# notebookservice-parent/notebookservice-web/notebookservice-swagger.yaml
swagger: '2.0'
info:
  description: >-
  version: 0.0.1-SNAPSHOT
  title: Notebook Service
basePath: /v2
...
schemes:
  - http
paths:
...
  /brand/{id}:
    get:
      tags:
        - Brand
      summary: Find brand by ID
      description: Returns a single brand
      operationId: getBrandById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of brand to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Brand'
...
  /modification/{id}:
    get:
      tags:
        - Modifications
      summary: Find modification by ID
      description: Returns a single modification
      operationId: getModificationById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of modification to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Modification'
...
  /client/{id}:
    get:
      tags:
        - Client
      summary: Find client by ID
      description: Returns a single client
      operationId: getClientById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of client to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Client'
...
  /product/{id}:
    get:
      tags:
        - Product
      summary: Find product by ID
      description: Returns a single product
      operationId: getProductById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of product to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Product'
...
  /maintenance/{id}:
    get:
      tags:
        - Maintenance
      summary: Find maintenance by ID
      description: Returns a single maintenance
      operationId: getMaintenanceById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of maintenance to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Maintenance'
...
definitions:
  Brand:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  Client:
    type: object
    properties:
      id:
        type: integer
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      phone:
        type: string
  Maintenance:
    type: object
    properties:
      id:
        type: integer
      startdate:
        type: string
        format: date
      enddate:
        type: string
        format: date
      status:
        type: string
        enum:
          - RECORDED
          - IN_PROGRESS
          - DONE
      fault:
        type: string
      user:
        $ref: '#/definitions/User'
      product:
        $ref: '#/definitions/Product'
      modification:
        type: array
        items:
          $ref: '#/definitions/Modification'
  Modification:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
  Product:
    type: object
    properties:
      id:
        type: integer
      description:
        type: string
      type:
        type: string
      brand:
        $ref: '#/definitions/Brand'
      client:
        $ref: '#/definitions/Client'
  User:
    type: object
    properties:
      id:
        type: integer
      username:
        type: string
      password:
        type: string
      userrole:
        type: string
        enum:
          - MANAGER
          - SERVICE